Bug#89789: Option to disable SSL/TLS on running server
Adding ssl_announce for announcing CLIENT_SSL capability
and also (dis)allowing new SSL connections.

Note:
If this is on and require_secure_transport is also on,
then it is only possible to log in through unix domain socket
or shared memory or other non-network transport layers.
For users that has 'REQUIRE SSL', then they cannot log in.

Also not tested with mysqlx protocol!